erikt: De-malloced code.  All arrays exist on stack.  I've also promioted exery float to double.

+++ src/utils.h
@@ -30,7 +30,7 @@ void free_matrix(double **x, int ysize);
 
 void printPhotom(Photometry * photom); 
 
-void doCentroid(float *pixels);
+void doCentroid(double *pixels);
 
 typedef int bool;
 #define true 1
